Refactor Gradio UI into modular webui/ package
- Extract each tab into independent component under webui/tabs/:
  remove_bg.py, boundary_crop.py, smart_crop.py, tagging.py, rename.py
- Each module exposes create_tab() with self-contained UI + processing logic
- webui/app.py assembles tabs into the main Blocks application
- Root app.py reduced to thin entry point (logging + launch)
