+++ webui/__init__.py
@@ -1,0 +1,1 @@
+"""Gradio web UI package for anime character image preprocessing."""
